Reorganise local area variables
