--- policyengine_us/variables/input/county_fips.py
@@ -1,8 +1,0 @@
-from policyengine_us.model_api import *
-
-
-class county_fips(Variable):
-    value_type = int
-    entity = Household
-    definition_period = YEAR
-    documentation = "County FIPS code"
