+++ policyengine_us/variables/input/geography.py
@@ -142,3 +142,18 @@ class reported_slspc(Variable):
     unit = USD
     definition_period = YEAR
     hidden_input = True
+
+
+class county_fips(Variable):
+    value_type = int
+    entity = Household
+    definition_period = YEAR
+    documentation = "County FIPS code"
+
+
+class state_fips(Variable):
+    value_type = int
+    entity = Household
+    definition_period = YEAR
+    documentation = "State FIPS code"
+    default_value = 6
